Removed check mode from running in test cycle (#632)
Check mode quite frankly is a pain in the ass.  Having it run by default
without the ability to turn it off only causes pain.  We have written
a lot of real-world roles, and we end up disabling check mode most of
the time.

This PR disables check mode by default, but the user is free to add it
to their test sequence if they want it.